state: move PortalTrieDB to new file
